updated app folder structure
