+++ app/__init__.py
@@ -1,0 +1,20 @@
+from flask import Flask
+from flask_bootstrap import Bootstrap
+
+from config import config
+
+
+bootstrap = Bootstrap()
+
+
+def create_app(config_name):
+    app = Flask(__name__)
+    app.config.from_object(config[config_name])
+    config[config_name].init_app(app)
+
+    bootstrap.init_app(app)
+
+    from app.views import bucket
+    app.register_blueprint(bucket)
+
+    return app
