+++ app/templates/home.html
@@ -1,0 +1,8 @@
+{% extends "base.html" %}
+{% block title %}Bucketlist Home{% endblock %}
+
+{% block page_content %}
+<div class="page-header">
+    <h1>Home</h1>
+</div>
+{% endblock %}
